+++ tests/MapOptionsNormalizerTest.php
@@ -2,8 +2,16 @@
 
 declare(strict_types=1);
 
-namespace Symfony\UX\Map\Tests;
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
+namespace Symfony\UX\Map\Tests;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\UX\Map\Exception\UnableToDenormalizeOptionsException;